+++ Path/DirectoryAnalyzers/EmptyAnalyzer.cs
@@ -1,0 +1,11 @@
+﻿namespace PathCli.DirectoryAnalyzers;
+
+public class EmptyAnalyzer : IDirectoryAnalyzer
+{
+    public PathProblem? Analyze(DirectoryInfo directory)
+    {
+        return directory.EnumerateDirectories().Any() || directory.EnumerateFiles().Any()
+            ? null
+            : PathProblem.Empty;
+    }
+}