+++ Path/DirectoryAnalyzers/ExistenceAnalyzer.cs
@@ -1,0 +1,9 @@
+﻿namespace PathCli.DirectoryAnalyzers;
+
+public class ExistenceAnalyzer : IDirectoryAnalyzer
+{
+    public PathProblem? Analyze(DirectoryInfo directory)
+    {
+        return directory.Exists ? null : PathProblem.Missing;
+    }
+}