+++ Path/DirectoryAnalyzers/IDirectoryAnalyzer.cs
@@ -1,0 +1,6 @@
+﻿namespace PathCli.DirectoryAnalyzers;
+
+public interface IDirectoryAnalyzer
+{
+    PathProblem? Analyze(DirectoryInfo directory);
+}