+++ Path/PathAnalyzer.cs
@@ -1,35 +1,37 @@
-﻿namespace PathCli;
+﻿using PathCli.DirectoryAnalyzers;
 
-/// <summary>
-/// Create an instance of PathAnalyzer.
-/// </summary>
-/// <param name="executableExtensions">List of executable extenions each prefixed with ".".</param>
-class PathAnalyzer(HashSet<string> executableExtensions)
+namespace PathCli;
+
+class PathAnalyzer(List<IDirectoryAnalyzer> directoryAnalyzers)
 {
-    public SortedDictionary<string, PathProblem>Analyze(PathString path)
+    public SortedDictionary<string, PathProblem> Analyze(PathString path)
     {
-        // basic analysis
-        var results = new SortedDictionary<string, PathProblem>(StringComparer.OrdinalIgnoreCase);
+        SortedDictionary<string, PathProblem> results = new(StringComparer.OrdinalIgnoreCase);
         foreach (string dir in path.Items)
         {
-            var problem = analyzeBasicIssues(dir);
-            if (problem != PathProblem.None)
+            DirectoryInfo dirInfo = new(dir);
+            var result = runDirectoryAnalyzers(dirInfo);
+            if (result is PathProblem problem)
             {
                 setProblem(results, dir, problem);
             }
         }
 
-        // analyze dupes
-        if (executableExtensions.Count != 0)
-        {
-            var dupes = path.Items.GroupBy(s => s).Where(g => g.Count() > 1).Select(g => g.Key).Distinct();
-            foreach (string dir in dupes)
-            {
-                setProblem(results, dir, PathProblem.Duplicate);
-            }
-        }
+        analyzeDupes(path, results);
 
         return results;
+    }
+
+    private static void analyzeDupes(PathString path, SortedDictionary<string, PathProblem> results)
+    {
+        var dupes = path.Items.GroupBy(s => s)
+            .Where(g => g.Count() > 1)
+            .Select(g => g.Key)
+            .Distinct();
+        foreach (string dir in dupes)
+        {
+            setProblem(results, dir, PathProblem.Duplicate);
+        }
     }
 
     private static void setProblem(SortedDictionary<string, PathProblem> results, string dir, PathProblem problem)
@@ -42,40 +44,17 @@ class PathAnalyzer(HashSet<string> executableExtensions)
         results.Add(dir, problem);
     }
 
-    private PathProblem analyzeBasicIssues(string dir)
+    private PathProblem? runDirectoryAnalyzers(DirectoryInfo dirInfo)
     {
-        DirectoryInfo dirInfo = new(dir);
-        var problem = checkExistence(dirInfo);
-        if (problem != PathProblem.None)
+        foreach (var analyzer in directoryAnalyzers)
         {
-            return problem;
-        }
-
-        problem = checkEmptyPath(dirInfo);
-        return problem != PathProblem.None ? problem : checkMissingExecutables(dirInfo);
-    }
-
-    private PathProblem checkMissingExecutables(DirectoryInfo dirInfo)
-    {
-        bool exeFound = false;
-        foreach (var file in dirInfo.EnumerateFiles())
-        {
-            if (executableExtensions.Contains(file.Extension, StringComparer.OrdinalIgnoreCase))
+            if (analyzer.Analyze(dirInfo) is PathProblem problem)
             {
-                exeFound = true;
-                break;
+                return problem;
             }
         }
-        return exeFound ? PathProblem.None : PathProblem.NoExecutables;
+
+        return null;
     }
 
-    private static PathProblem checkEmptyPath(DirectoryInfo dirInfo)
-    {
-        return dirInfo.EnumerateDirectories().Any() ? PathProblem.None : PathProblem.Empty;
-    }
-
-    private static PathProblem checkExistence(DirectoryInfo dirInfo)
-    {
-        return !dirInfo.Exists ? PathProblem.Missing : PathProblem.None;
-    }
 }